Funciona no lo toque

+++ lora_receiver/src/services.h
@@ -3,7 +3,7 @@
 
 void Lora_connection();
 void WiFi_connection();
-DynamicJsonDocument Create_orion_package(String message, int rssi, float snr, int error_hz);
+DynamicJsonDocument Create_orion_package(String message, int rssi, float snr, long error_hz);
 bool Has_description_and_type(const DynamicJsonDocument &doc, const char *wanted_description, const char *wanted_type);
 
 #endif
